al menos las balas se ven bien

+++ project/src/our_scripts/components/Weapon.cpp
@@ -7,7 +7,7 @@
 
 //Weapon::Weapon() : _damage(0), _cooldown(0.0f), _distance(0.0f), _speed(0.0f), _tex(""), _attack_width(50), _attack_height(50) {}
 
-Weapon::Weapon(int d, float cd, float dis, float s, const std::string& t, int w, int h)
+Weapon::Weapon(int d, float cd, float dis, float s, const std::string& t, float w, float h)
 	: _damage(d), _cooldown(cd), _distance(dis), _speed(s), _tex(t), 
 	_attack_width(w), _attack_height(h), _lastShoot(0.0f), _tr(nullptr) 
 {
@@ -20,7 +20,7 @@ Weapon::initComponent() {
 Weapon::~Weapon() {}
 
 void 
-Weapon::set_attack_size(int w, int h) {
+Weapon::set_attack_size(float w, float h) {
 	_attack_height = h; 
 	_attack_width = w;
 }
